refactor: rename core -> internal/utils
