+++ src/internal/utils/data-structure.ts
