+++ src/internal/utils/file-system.ts
