+++ src/internal/utils/ts/diagnostics.ts
