+++ src/internal/utils/xml-string.ts
