+++ src/tsconfig/resolver.ts
@@ -3,8 +3,8 @@
  */
 
 import { join, resolve } from "node:path";
-import { fileExists, isDirectory } from "../core/files.ts";
 import { TsInspectError } from "../error.ts";
+import { fileExists, isDirectory } from "../internal/utils/file-system.ts";
 
 /**
  * Resolves a project path. Accepts either:
